moved life statements API handler to question domanin
