+++ question/life_statements.go
@@ -1,4 +1,4 @@
-package rules
+package question
 
 import (
 	"encoding/json"
@@ -28,22 +28,19 @@ func LifeStatements(w http.ResponseWriter, r *http.Request) (string, interface{}
 	policyJson, err := policy.Marshal()
 	lib.CheckError(err)
 
+	fx := new(models.Fx)
+
 	statements := &Statements{
 		Statements: make([]*models.Statement, 0),
 		Text:       "",
 	}
 
-	rulesFile = getRulesFile(grule, rulesFilename)
+	rulesFile = lib.GetRulesFile(grule, rulesFilename)
 
-	_, ruleOutput := rulesFromJson(rulesFile, statements, policyJson, nil)
+	_, ruleOutput := lib.RulesFromJsonV2(fx, rulesFile, statements, policyJson, nil)
 
 	ruleOutputJson, err := json.Marshal(ruleOutput)
 	lib.CheckError(err)
 
 	return string(ruleOutputJson), ruleOutput, nil
-}
-
-type Statements struct {
-	Statements []*models.Statement `json:"statements"`
-	Text       string              `json:"text,omitempty"`
 }
